refactor: move main loop task into separate function
