+++ src/plugins/skills/firemaking/index.ts
@@ -4,6 +4,7 @@ import { loopingEvent } from '@engine/plugins';
 import { FIREMAKING_LOGS } from './data';
 import { canChain, canLight } from './chance';
 import { lightFire } from './light-fire';
+import { runFiremakingTask } from './firemaking-task';
 
 const action: itemOnItemActionHandler = (details) => {
     const { player, usedItem, usedWithItem, usedSlot, usedWithSlot } = details;
@@ -15,7 +16,6 @@ const action: itemOnItemActionHandler = (details) => {
     const log = usedItem.itemId !== itemIds.tinderbox ? usedItem : usedWithItem;
     const removeFromSlot = usedItem.itemId !== itemIds.tinderbox ? usedSlot : usedWithSlot;
     const skillInfo = FIREMAKING_LOGS.find(l => l.logItem.gameId === log.itemId);
-    const position = player.position;
 
     if(!skillInfo) {
         player.sendMessage(`Mishandled firemaking log ${log.itemId}.`);
@@ -30,43 +30,11 @@ const action: itemOnItemActionHandler = (details) => {
 
     if(player.metadata.lastFire && Date.now() - player.metadata.lastFire < 1200 &&
         canChain(skillInfo.requiredLevel, player.skills.firemaking.level)) {
-        lightFire(player, position, worldItemLog, skillInfo.experienceGained);
+        lightFire(player, player.position, worldItemLog, skillInfo.experienceGained);
     } else {
         player.sendMessage(`You attempt to light the logs.`);
 
-        let canLightFire = false;
-        let elapsedTicks = 0;
-        const loop = loopingEvent({ player });
-        loop.event.subscribe(() => {
-            if(worldItemLog.removed) {
-                loop.cancel();
-                return;
-            }
-
-            if(canLightFire) {
-                loop.cancel();
-                player.busy = true;
-                setTimeout(() => lightFire(player, position, worldItemLog, skillInfo.experienceGained), 1200);
-                return;
-            }
-
-            // @TODO check for existing location objects again (in-case one spawned here during this loop)
-            // @TODO check for tinderbox in-case it was removed
-
-            if(elapsedTicks === 0 || elapsedTicks % 12 === 0) {
-                player.playAnimation(animationIds.lightingFire);
-            }
-
-            canLightFire = elapsedTicks > 10 && canLight(skillInfo.requiredLevel, player.skills.firemaking.level);
-
-            if(!canLightFire && (elapsedTicks === 0 || elapsedTicks % 4 === 0)) {
-                player.playSound(soundIds.lightingFire, 10, 0);
-            } else if(canLightFire) {
-                player.playSound(soundIds.fireLit, 7);
-            }
-
-            elapsedTicks++;
-        });
+        runFiremakingTask(player, worldItemLog, skillInfo);
     }
 };
